restore unminify module scripts, move and cleanup more code.
